+++ dj3base/tests/utils.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+import sys
+
+
+_IS_UNIT_TESTING = sys.argv[1:2] == ['test']
+
+
+def is_unit_testing() -> bool:
+    return _IS_UNIT_TESTING