+++ HISTORY.md
@@ -1,5 +1,10 @@
 # History
 
+## 0.1.1
+
+1. add commands.SingletonCommand
+2. add tests.utils.is_unit_testing
+
 ## 0.1.0
 
 1. add type_hints